// leoz-web/src/app/core/models/importscan.model.ts
export interface Importscan {
  senderline1: string;
  deliveryline1: string;
  senddate: string;
  senderstation: number;
  deliverystation: number;
  deliveryzip: string;
  deliverycity: string;
  parcelno: string;
  deliverydate: string;
  deliverytime: string;
  typeOfPackaging: number;
  realWeight: number;
  noOfPacks: number;
  volWeight: number;
  length: number;
  width: number;
  height: number
}